Move unitary_tools.py to simulation.tools
