+++ pyquil/tests/test_reference_density_simulator.py
@@ -12,8 +12,8 @@ from pyquil.operator_estimation import measure_observables
 from pyquil.paulis import sI, sX, sY, sZ
 from pyquil.pyqvm import PyQVM
 from pyquil.reference_simulator import ReferenceDensitySimulator, _is_valid_quantum_state
+from pyquil.simulation.tools import lifted_gate_matrix
 from pyquil.tests.utils import DummyCompiler
-from pyquil.unitary_tools import lifted_gate_matrix
 
 
 def test_qaoa_density():
